refactor(notifications): split event manager helpers
Move notification state adapters, discussion text parsing, HWO/SPS product workflows, and SPS matching into focused helper modules while keeping the manager's compatibility surface intact.

Tested: uv run pytest -q -n 0 --tb=short tests/test_notification_event_manager.py tests/test_notification_hwo_update.py tests/test_notification_sps_dedupe.py tests/test_notification_state_hwo_sps_fields.py tests/test_runtime_state_hwo_sps.py tests/test_split_notification_timers.py tests/test_split_update_timers.py; uv run ruff check src tests scripts; uv run pyright

Confidence: high
