+++ src/accessiweather/notifications/notification_event_manager.py
@@ -12,12 +12,9 @@ enabled in Settings > Notifications.
 
 from __future__ import annotations
 
-import hashlib
 import json
 import logging
-import re
-from dataclasses import dataclass, field
-from datetime import UTC, datetime, timedelta
+from datetime import datetime
 from pathlib import Path
 from typing import TYPE_CHECKING
 
@@ -28,6 +25,27 @@ from .minutely_precipitation import (
     build_minutely_transition_signature,
     detect_minutely_precipitation_likelihood,
     detect_minutely_precipitation_transition,
+)
+from .notification_event_products import check_hwo_update, check_sps_new
+from .notification_event_state import (
+    NotificationEvent,
+    NotificationState,
+    legacy_shape_to_runtime_section,
+    runtime_section_to_legacy_shape,
+)
+from .notification_event_text import (
+    extract_discussion_issued_time_label as _extract_discussion_issued_time_label,
+    extract_section as _extract_section,
+    format_issuance_time_label as _format_issuance_time_label,
+    get_risk_category,
+    hash_product_text,
+    is_no_change_summary as _is_no_change_summary,
+    summarize_discussion_change,
+)
+from .notification_sps_helpers import (
+    format_sps_body,
+    sps_alert_signatures,
+    sps_is_case_a,
 )
 
 if TYPE_CHECKING:
@@ -43,265 +61,7 @@ if TYPE_CHECKING:
     )
 
 
-# Rate-limit window for per-(product, location) notifications. The HWO check
-# suppresses re-dispatch within this window while still advancing stored state
-# so we don't re-trigger on the next cycle forever.
-_HWO_RATE_LIMIT_WINDOW = timedelta(minutes=30)
-# SPS (Special Weather Statement) reuses the same 30-minute sliding window,
-# keyed independently on ``self._last_product_notified_at``.
-_SPS_RATE_LIMIT_WINDOW = timedelta(minutes=30)
-# Windows toast bodies clip around 200 chars — be safely under that.
-_SPS_MAX_BODY_CHARS = 160
-# Summaries shorter than this (after stripping) fall back to the generic body.
-_HWO_SUMMARY_MIN_CHARS = 20
-
 logger = logging.getLogger(__name__)
-
-
-def _extract_discussion_issued_time_label(discussion_text: str | None) -> str | None:
-    """Extract the station-local issued time label from AFD text when present."""
-    if not discussion_text:
-        return None
-
-    patterns = (
-        r"\bISSUED\s+(\d{1,2})(\d{2})\s+([AP]M)\s+([A-Z]{2,4})\b",
-        r"\bISSUED\s+(\d{1,2}):(\d{2})\s+([AP]M)\s+([A-Z]{2,4})\b",
-        r"(?:^|\n)\s*(\d{1,2})(\d{2})\s+([AP]M)\s+([A-Z]{2,4})\s+(?:Mon|Tue|Wed|Thu|Fri|Sat|Sun)\b",
-        r"(?:^|\n)\s*(\d{1,2}):(\d{2})\s+([AP]M)\s+([A-Z]{2,4})\s+(?:Mon|Tue|Wed|Thu|Fri|Sat|Sun)\b",
-    )
-
-    for pattern in patterns:
-        match = re.search(pattern, discussion_text, re.IGNORECASE)
-        if match:
-            hour, minute, meridiem, tz_name = match.groups()
-            return f"{int(hour)}:{minute} {meridiem.upper()} {tz_name.upper()}"
-
-    return None
-
-
-def _format_issuance_time_label(issuance_time: datetime) -> str:
-    """Format issuance time using the datetime's own timezone context."""
-    time_str = issuance_time.strftime("%I:%M %p").lstrip("0")
-    tz_name = issuance_time.tzname() or ""
-    return f"{time_str} {tz_name}".strip()
-
-
-def get_risk_category(risk: int) -> str:
-    """
-    Categorize severe weather risk level.
-
-    Uses the same thresholds as the UI display in current_conditions.py.
-
-    Args:
-        risk: Risk percentage (0-100)
-
-    Returns:
-        Category name: 'minimal', 'low', 'moderate', 'high', or 'extreme'
-
-    """
-    if risk >= 80:
-        return "extreme"
-    if risk >= 60:
-        return "high"
-    if risk >= 40:
-        return "moderate"
-    if risk >= 20:
-        return "low"
-    return "minimal"
-
-
-def _extract_section(text: str, start_marker: str, end_markers: tuple[str, ...]) -> str | None:
-    """
-    Extract the content of a named section from AFD text.
-
-    Args:
-        text: The full AFD text to search.
-        start_marker: The section header to look for (e.g. '.WHAT HAS CHANGED...').
-        end_markers: Tuple of prefixes that indicate the end of the section
-                     (e.g. lines starting with '.' or '&&').
-
-    Returns:
-        The extracted section body (stripped), or None if the section is not found.
-
-    """
-    lines = text.splitlines()
-    in_section = False
-    body_lines: list[str] = []
-    for line in lines:
-        stripped = line.strip()
-        if not in_section:
-            if stripped.upper().startswith(start_marker.upper()):
-                in_section = True
-            continue
-        # We are inside the section — check for terminator
-        if any(stripped.startswith(m) for m in end_markers) or any(
-            stripped.upper().startswith(m.upper()) for m in end_markers
-        ):
-            break
-        body_lines.append(stripped)
-    if not in_section:
-        return None
-    content = " ".join(part for part in body_lines if part)
-    return content if content else None
-
-
-def _normalize_discussion_summary(text: str | None) -> str:
-    """Normalize extracted AFD summary text for comparison."""
-    if not text:
-        return ""
-    return " ".join(text.split()).casefold().strip(" .")
-
-
-def _is_no_change_summary(text: str | None) -> bool:
-    """Return True when a WHAT HAS CHANGED section explicitly says nothing changed."""
-    normalized = _normalize_discussion_summary(text)
-    if not normalized:
-        return False
-    return normalized in {
-        "no change",
-        "no changes",
-        "no significant change",
-        "no significant changes",
-        "no significant changes made to forecast",
-        "no significant changes made to the forecast",
-    }
-
-
-def summarize_discussion_change(previous_text: str | None, current_text: str | None) -> str | None:
-    """
-    Return a short human-friendly summary of what changed in the discussion text.
-
-    Priority order:
-    1. `.WHAT HAS CHANGED...` section — extract body up to the next section marker.
-    2. `.KEY MESSAGES...` section — extract body up to ``&&``.
-    3. First new line not present in the previous discussion text.
-
-    The result is truncated to ~300 characters.
-    """
-    if not current_text:
-        return None
-
-    # 1. Try .WHAT HAS CHANGED... section. If it explicitly says there were no
-    # changes, keep looking for changed key messages before giving up.
-    what_changed_section = _extract_section(
-        current_text,
-        start_marker=".WHAT HAS CHANGED",
-        end_markers=(".", "&&"),
-    )
-    current_declares_no_changes = _is_no_change_summary(what_changed_section)
-    if what_changed_section and not current_declares_no_changes:
-        return what_changed_section[:300]
-
-    # 2. Try .KEY MESSAGES... section
-    section = _extract_section(
-        current_text,
-        start_marker=".KEY MESSAGES",
-        end_markers=(".", "&&"),
-    )
-    if section:
-        previous_section = _extract_section(
-            previous_text or "",
-            start_marker=".KEY MESSAGES",
-            end_markers=(".", "&&"),
-        )
-        if not previous_section or _normalize_discussion_summary(
-            section
-        ) != _normalize_discussion_summary(previous_section):
-            return section[:300]
-        if current_declares_no_changes:
-            return None
-
-    if current_declares_no_changes:
-        return None
-
-    # 3. Fall back to first new line not present in previous text
-    previous_lines = {
-        line.strip()
-        for line in (previous_text or "").splitlines()
-        if line.strip() and not line.strip().startswith("$")
-    }
-    for raw_line in current_text.splitlines():
-        line = raw_line.strip()
-        if not line or line.startswith("$"):
-            continue
-        if line not in previous_lines:
-            return line[:300]
-    return None
-
-
-@dataclass
-class NotificationEvent:
-    """Represents a notification event to be sent."""
-
-    event_type: str
-    title: str
-    message: str
-    sound_event: str  # Sound event key for the soundpack
-
-
-@dataclass
-class NotificationState:
-    """Tracks state for notification change detection."""
-
-    last_discussion_issuance_time: datetime | None = None  # NWS API issuanceTime
-    last_discussion_text: str | None = None
-    last_severe_risk: int | None = None
-    last_minutely_transition_signature: str | None = None
-    last_minutely_likelihood_signature: str | None = None
-    last_check_time: datetime | None = None
-    # HWO (Hazardous Weather Outlook) tracking — Unit 10 populates these.
-    last_hwo_issuance_time: datetime | None = None
-    last_hwo_text: str | None = None
-    last_hwo_summary_signature: str | None = None
-    # SPS (Special Weather Statement) tracking — Unit 11 populates these.
-    # Stored as a set for O(1) dedupe; round-trips through JSON as sorted list.
-    last_sps_product_ids: set[str] = field(default_factory=set)
-
-    def to_dict(self) -> dict:
-        """Convert to dictionary for persistence."""
-        return {
-            "last_discussion_issuance_time": (
-                self.last_discussion_issuance_time.isoformat()
-                if self.last_discussion_issuance_time
-                else None
-            ),
-            "last_discussion_text": self.last_discussion_text,
-            "last_severe_risk": self.last_severe_risk,
-            "last_minutely_transition_signature": self.last_minutely_transition_signature,
-            "last_minutely_likelihood_signature": self.last_minutely_likelihood_signature,
-            "last_check_time": self.last_check_time.isoformat() if self.last_check_time else None,
-            "last_hwo_issuance_time": (
-                self.last_hwo_issuance_time.isoformat() if self.last_hwo_issuance_time else None
-            ),
-            "last_hwo_text": self.last_hwo_text,
-            "last_hwo_summary_signature": self.last_hwo_summary_signature,
-            # Sort for deterministic output (JSON diffs, snapshot tests).
-            "last_sps_product_ids": sorted(self.last_sps_product_ids),
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict) -> NotificationState:
-        """Create from dictionary."""
-        last_check = data.get("last_check_time")
-        last_issuance = data.get("last_discussion_issuance_time")
-        last_hwo_issuance = data.get("last_hwo_issuance_time")
-        sps_ids = data.get("last_sps_product_ids") or []
-        return cls(
-            last_discussion_issuance_time=(
-                datetime.fromisoformat(last_issuance) if last_issuance else None
-            ),
-            last_discussion_text=data.get("last_discussion_text"),
-            last_severe_risk=data.get("last_severe_risk"),
-            last_minutely_transition_signature=data.get("last_minutely_transition_signature"),
-            last_minutely_likelihood_signature=data.get("last_minutely_likelihood_signature"),
-            last_check_time=datetime.fromisoformat(last_check) if last_check else None,
-            last_hwo_issuance_time=(
-                datetime.fromisoformat(last_hwo_issuance) if last_hwo_issuance else None
-            ),
-            last_hwo_text=data.get("last_hwo_text"),
-            last_hwo_summary_signature=data.get("last_hwo_summary_signature"),
-            last_sps_product_ids=set(sps_ids),
-        )
 
 
 class NotificationEventManager:
@@ -383,63 +143,12 @@ class NotificationEventManager:
     @staticmethod
     def _runtime_section_to_legacy_shape(section: dict) -> dict:
         """Convert unified runtime state to the legacy notification-state shape."""
-        discussion = section.get("discussion", {})
-        severe_risk = section.get("severe_risk", {})
-        minutely_precipitation = section.get("minutely_precipitation", {})
-        hwo = section.get("hwo", {})
-        sps = section.get("sps", {})
-        return {
-            "last_discussion_issuance_time": discussion.get("last_issuance_time"),
-            "last_discussion_text": discussion.get("last_text"),
-            "last_severe_risk": severe_risk.get("last_value"),
-            "last_minutely_transition_signature": minutely_precipitation.get(
-                "last_transition_signature"
-            ),
-            "last_minutely_likelihood_signature": minutely_precipitation.get(
-                "last_likelihood_signature"
-            ),
-            "last_check_time": discussion.get("last_check_time")
-            or severe_risk.get("last_check_time")
-            or minutely_precipitation.get("last_check_time"),
-            "last_hwo_issuance_time": hwo.get("last_issuance_time"),
-            "last_hwo_text": hwo.get("last_text"),
-            "last_hwo_summary_signature": hwo.get("last_summary_signature"),
-            "last_sps_product_ids": list(sps.get("last_product_ids") or []),
-        }
+        return runtime_section_to_legacy_shape(section)
 
     @staticmethod
     def _legacy_shape_to_runtime_section(data: dict) -> dict:
         """Convert legacy notification-state payloads to the unified section shape."""
-        last_check_time = data.get("last_check_time")
-        sps_ids = data.get("last_sps_product_ids") or []
-        return {
-            "discussion": {
-                "last_issuance_time": data.get("last_discussion_issuance_time"),
-                "last_text": data.get("last_discussion_text"),
-                "last_check_time": last_check_time,
-            },
-            "severe_risk": {
-                "last_value": data.get("last_severe_risk"),
-                "last_check_time": last_check_time,
-            },
-            "minutely_precipitation": {
-                "last_transition_signature": data.get("last_minutely_transition_signature"),
-                "last_likelihood_signature": data.get("last_minutely_likelihood_signature"),
-                "last_check_time": last_check_time,
-            },
-            "hwo": {
-                "last_issuance_time": data.get("last_hwo_issuance_time"),
-                "last_text": data.get("last_hwo_text"),
-                "last_summary_signature": data.get("last_hwo_summary_signature"),
-                "last_check_time": last_check_time,
-            },
-            "sps": {
-                # Sorted so round-trips are stable; NotificationState's
-                # ``set`` field re-materializes order-independence.
-                "last_product_ids": sorted(sps_ids),
-                "last_check_time": last_check_time,
-            },
-        }
+        return legacy_shape_to_runtime_section(data)
 
     def check_for_events(
         self,
@@ -604,105 +313,21 @@ class NotificationEventManager:
         hwo_product: TextProduct | None,
         settings: AppSettings,
     ) -> None:
-        """
-        Inspect a freshly-fetched HWO product and notify on material updates.
-
-        Unit 10 — Hazardous Weather Outlook change detection. The event loop
-        feeds us the pre-warmed HWO for ``location``; we:
-
-        1. Seed baseline state on the first fetch (no notification).
-        2. Compare ``issuance_time`` + content-signature against stored values;
-           short-circuit when both are unchanged.
-        3. On change, persist the new baseline, then — subject to the
-           ``notify_hwo_update`` setting (default True when absent) and a
-           sliding 30-minute rate-limit bucket keyed by ``("HWO", location)`` —
-           dispatch a desktop notification.
-
-        ``None`` products and locations without a ``cwa_office`` are no-ops.
-        """
-        if hwo_product is None:
-            return
-        if not getattr(location, "cwa_office", None):
-            return
-
-        new_issuance = hwo_product.issuance_time
-        new_text = hwo_product.product_text or ""
-        signature = self._hash_product_text(new_text)
-
-        stored_issuance = self.state.last_hwo_issuance_time
-        stored_signature = self.state.last_hwo_summary_signature
-        stored_text = self.state.last_hwo_text
-
-        # Cold-start: no stored baseline → record silently, never dispatch.
-        if stored_issuance is None and stored_signature is None:
-            self.state.last_hwo_issuance_time = new_issuance
-            self.state.last_hwo_text = new_text
-            self.state.last_hwo_summary_signature = signature
-            self._save_state()
-            logger.debug(
-                "_check_hwo_update: first-run baseline for %s (%s) — no notification",
-                location.name,
-                location.cwa_office,
-            )
-            return
-
-        # Unchanged → true no-op: no state churn, no dispatch.
-        if stored_issuance == new_issuance and stored_signature == signature:
-            return
-
-        # Changed: persist new baseline before deciding whether to notify so we
-        # never re-fire against the old state if dispatch is suppressed below.
-        self.state.last_hwo_issuance_time = new_issuance
-        self.state.last_hwo_text = new_text
-        self.state.last_hwo_summary_signature = signature
-        self._save_state()
-
-        if not getattr(settings, "notify_hwo_update", True):
-            logger.debug(
-                "_check_hwo_update: notify_hwo_update disabled — suppressing dispatch for %s",
-                location.name,
-            )
-            return
-
-        bucket = ("HWO", location.name)
-        now = datetime.now(UTC)
-        last_sent = self._last_product_notified_at.get(bucket)
-        if last_sent is not None and now - last_sent < _HWO_RATE_LIMIT_WINDOW:
-            logger.debug(
-                "_check_hwo_update: rate-limited for %s (last=%s) — state updated, no dispatch",
-                location.name,
-                last_sent,
-            )
-            return
-        self._last_product_notified_at[bucket] = now
-
-        message = self._format_hwo_body(stored_text, hwo_product)
-        logger.info(
-            "HWO updated for %s (%s): %s -> %s",
-            location.name,
-            location.cwa_office,
-            stored_issuance,
-            new_issuance,
-        )
-        self._dispatch_hwo_notification(
-            location=location,
-            product=hwo_product,
-            message=message,
-        )
+        """Inspect a freshly fetched HWO product and notify on material updates."""
+        check_hwo_update(self, location, hwo_product, settings)
 
     @staticmethod
     def _hash_product_text(text: str) -> str:
         """Return a stable signature for an HWO product text."""
-        normalized = (text or "").strip()
-        return hashlib.sha256(normalized.encode("utf-8", errors="replace")).hexdigest()
+        return hash_product_text(text)
 
     @staticmethod
     def _format_hwo_body(stored_text: str | None, new_product: TextProduct) -> str:
         """Produce the notification body — prefer summarizer output, fall back to generic."""
         summary = summarize_discussion_change(stored_text, new_product.product_text)
-        if summary and len(summary.strip()) > _HWO_SUMMARY_MIN_CHARS:
+        if summary and len(summary.strip()) > 20:
             return summary.strip()
-        return f"Hazardous Weather Outlook updated for {new_product.cwa_office} — tap to view."
+        return f"Hazardous Weather Outlook updated for {new_product.cwa_office} - tap to view."
 
     def _dispatch_hwo_notification(
         self,
@@ -742,213 +367,23 @@ class NotificationEventManager:
         cached_alerts: Sequence[WeatherAlert] | None,
         settings: AppSettings,
     ) -> None:
-        """
-        Dispatch notifications for informational SPS products only.
-
-        NWS issues two populations of Special Weather Statements:
-
-        - **Case A — event-style.** Hail, dense fog, strong-thunderstorm
-          statements that also appear on ``/alerts/active``. The existing alert
-          pipeline already surfaces them; notifying again here would
-          double-notify. We detect these via headline/first-description
-          substring matching against the SPS product text.
-        - **Case B — informational.** Fire-weather, pollen, coordination
-          statements that appear on ``/products/types/SPS`` but never on
-          ``/alerts/active``. These are invisible to users today; we notify on
-          them.
-
-        The heuristic is intentionally conservative: when a Case A match is not
-        obvious we favor notifying (honoring the plan's "if in doubt, show the
-        informational SPS the user would otherwise miss" principle).
-
-        Behavior summary:
-          1. Cold start — first ever fetch for this manager records all IDs
-             silently. Detected by ``not last_sps_product_ids`` and the bucket
-             not yet present in ``_sps_cold_started``.
-          2. Expiration — IDs in stored state but absent from the fetch are
-             silently removed.
-          3. For each NEW ID (not yet in state), classify Case A vs Case B;
-             add to state unconditionally; dispatch only for Case B, subject
-             to the setting toggle and the 30-min rate-limit bucket.
-          4. ``None`` or empty product list after cold-start just expires
-             stored IDs.
-        """
-        products = list(sps_products or [])
-        alerts = list(cached_alerts or [])
-
-        # Cold-start: first call ever for this (manager, location) pair.
-        if not hasattr(self, "_sps_cold_started"):
-            self._sps_cold_started: set[str] = set()  # type: ignore[attr-defined]
-
-        bucket_key = ("SPS", location.name)
-        is_cold_start = (
-            location.name not in self._sps_cold_started and not self.state.last_sps_product_ids
-        )
-
-        current_ids = {p.product_id for p in products}
-
-        if is_cold_start:
-            if current_ids:
-                self.state.last_sps_product_ids |= current_ids
-                self._save_state()
-            self._sps_cold_started.add(location.name)
-            logger.debug(
-                "_check_sps_new: cold-start baseline for %s (%d ids) — no dispatch",
-                location.name,
-                len(current_ids),
-            )
-            return
-
-        # Expire any stored IDs that are no longer present.
-        stale = self.state.last_sps_product_ids - current_ids
-        if stale:
-            self.state.last_sps_product_ids -= stale
-            self._save_state()
-            logger.debug("_check_sps_new: expired %d SPS id(s) for %s", len(stale), location.name)
-
-        # Evaluate only genuinely new product IDs.
-        new_products = [p for p in products if p.product_id not in self.state.last_sps_product_ids]
-        if not new_products:
-            return
-
-        # Build a normalized lookup of active SPS alert headlines/first-lines
-        # once for O(n + m) matching rather than O(n * m) per product.
-        alert_signatures = self._sps_alert_signatures(alerts)
-        enabled = getattr(settings, "notify_sps_issued", True)
-
-        dispatched_this_call = False
-        for product in new_products:
-            # Always add to seen state first — Case A or rate-limited products
-            # must never re-evaluate next cycle.
-            self.state.last_sps_product_ids.add(product.product_id)
-
-            if self._sps_is_case_a(product, alert_signatures):
-                logger.debug(
-                    "_check_sps_new: Case A (event-style) suppressed — product=%s",
-                    product.product_id,
-                )
-                continue
-
-            # Case B — informational.
-            if not enabled:
-                logger.debug(
-                    "_check_sps_new: notify_sps_issued disabled — suppressing %s",
-                    product.product_id,
-                )
-                continue
-
-            now = datetime.now(UTC)
-            last_sent = self._last_product_notified_at.get(bucket_key)
-            if last_sent is not None and now - last_sent < _SPS_RATE_LIMIT_WINDOW:
-                logger.debug(
-                    "_check_sps_new: rate-limited for %s (last=%s) — state updated, no dispatch",
-                    location.name,
-                    last_sent,
-                )
-                continue
-            if dispatched_this_call:
-                # Stamp once per call so repeated Case B products in the same
-                # fetch still fall under the rate limit — matches HWO shape.
-                logger.debug(
-                    "_check_sps_new: additional Case B in same fetch rate-limited — %s",
-                    product.product_id,
-                )
-                continue
-
-            self._last_product_notified_at[bucket_key] = now
-            dispatched_this_call = True
-            message = self._format_sps_body(product)
-            logger.info(
-                "SPS informational dispatch for %s (%s): %s",
-                location.name,
-                product.cwa_office,
-                product.product_id,
-            )
-            self._dispatch_sps_notification(
-                location=location,
-                product=product,
-                message=message,
-            )
-
-        self._save_state()
-
-    @staticmethod
-    def _normalize_for_match(text: str | None) -> str:
-        """Return a casefolded, whitespace-collapsed version of ``text`` for matching."""
-        if not text:
-            return ""
-        return " ".join(text.split()).casefold()
+        """Dispatch notifications for informational SPS products only."""
+        check_sps_new(self, location, sps_products, cached_alerts, settings)
 
     @classmethod
     def _sps_alert_signatures(cls, alerts: Sequence[WeatherAlert]) -> list[str]:
-        """
-        Collect normalized signatures for active SPS alerts.
-
-        Only alerts whose ``event`` is ``"Special Weather Statement"``
-        (case-insensitive) are considered. For each, we take the ``headline``
-        plus the first non-empty line of ``description``; both are normalized
-        via :meth:`_normalize_for_match`. Empty signatures are dropped.
-        """
-        signatures: list[str] = []
-        for alert in alerts:
-            event = (alert.event or "").strip().casefold()
-            if event != "special weather statement":
-                continue
-            for candidate in (alert.headline, cls._first_nonempty_line(alert.description)):
-                sig = cls._normalize_for_match(candidate)
-                if sig:
-                    signatures.append(sig)
-        return signatures
+        """Collect normalized signatures for active SPS alerts."""
+        return sps_alert_signatures(alerts)
 
     @classmethod
     def _sps_is_case_a(cls, product: TextProduct, alert_signatures: Sequence[str]) -> bool:
-        """
-        Return True when ``product`` looks like the event-style SPS an alert covers.
-
-        Matching rule: any active SPS alert's normalized headline or first
-        description line appears as a substring of the product's normalized
-        text (headline + product_text), OR vice versa. The bidirectional check
-        covers the rare case where an alert description runs longer than the
-        product blurb, even though the common shape is alert-headline ⊆
-        product-text.
-        """
-        if not alert_signatures:
-            return False
-        product_haystack_parts = [product.headline or "", product.product_text or ""]
-        product_norm = cls._normalize_for_match(" ".join(product_haystack_parts))
-        if not product_norm:
-            return False
-        return any(sig in product_norm or product_norm in sig for sig in alert_signatures)
+        """Return True when ``product`` looks like the event-style SPS an alert covers."""
+        return sps_is_case_a(product, alert_signatures)
 
     @classmethod
     def _format_sps_body(cls, product: TextProduct) -> str:
         """Build the toast body — headline + CWA office, with text fallback."""
-        headline = (product.headline or "").strip()
-        if headline:
-            body = f"{headline} \u2014 {product.cwa_office}"
-            return cls._truncate(body, _SPS_MAX_BODY_CHARS)
-        fallback = cls._first_nonempty_line(product.product_text) or ""
-        return cls._truncate(fallback.strip(), _SPS_MAX_BODY_CHARS)
-
-    @staticmethod
-    def _first_nonempty_line(text: str | None) -> str | None:
-        """Return the first non-empty, non-whitespace line of ``text``."""
-        if not text:
-            return None
-        for raw in text.splitlines():
-            line = raw.strip()
-            if line:
-                return line
-        return None
-
-    @staticmethod
-    def _truncate(text: str, limit: int) -> str:
-        """Truncate ``text`` to ``limit`` chars, appending an ellipsis if cut."""
-        if len(text) <= limit:
-            return text
-        if limit <= 3:
-            return text[:limit]
-        return text[: limit - 3] + "..."
+        return format_sps_body(product)
 
     def _dispatch_sps_notification(
         self,
